Cleanup/reorg
